feat: it reduces the dashboard view complexity
