+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './getMedicationsParamsReducer'
 export * from './useAuth'
 export * from './useCreateMedicine'
+export * from './useGetMedications'
 export * from './useQueryParams'
